Implement PUT

+++ models.go
@@ -4,7 +4,7 @@ import "github.com/guregu/null"
 
 // Note struct
 type Note struct {
-	ID      int64       `json:"id"`
+	ID      int         `json:"id"`
 	Title   string      `json:"title"`
 	Body    null.String `json:"body"`
 	Created string      `json:"created"`
@@ -13,7 +13,7 @@ type Note struct {
 
 // NoteTitle struct
 type NoteTitle struct {
-	ID      int64  `json:"id"`
+	ID      int    `json:"id"`
 	Title   string `json:"title"`
 	Created string `json:"created"`
 	Updated string `json:"updated"`
